requirements.txt
fastapi>=0.110.2
pydantic>=2.7.0
uvicorn>=0.29.0
jinja2>=3.1.3
numpy>=1.25.4
Pillow>=10.3.0
multipart>=0.0.9
matplotlib>=3.8.4